Add command line arguemnts

+++ internal/encoder/encoder_config.go
@@ -1,8 +1,8 @@
 package encoder
 
 type EncoderConfig struct {
-	FrameRate int
-	ImageWidth uint16 
-	ImageHeight uint16
-	MultipleFiles bool // If True, encodes all files in the directory. If False, then encodes the first file 
+	FrameRate     int
+	ImageWidth    uint
+	ImageHeight   uint
+	MultipleFiles bool // If True, encodes all files in the directory. If False, then encodes the first file
 }
